ready for testing

+++ myheaders.h
@@ -17,5 +17,5 @@ void 	heapSort(long data[], long n);
 void 	heapify(long data[], long n, long i);
 void 	selectionSort(long arr[], long n);
 void 	quickSort(long arr[], long n);
-void 	bubbleSort(long data[], int n);
+void 	bubbleSort(long data[], long n);
 void 	countingSort(long data[], const long n);
